fix: fix incorrect padding space
